# docs/openapi.yaml
openapi: 3.1.0
info:
  title: HiTechCloud AI OS API
  version: 0.1.0
  description: |
    AI Operating System — Unified Gateway for MCP, Skills, and Plugins.
    5-Layer Architecture: AG-UI → A2A → MCP → REST → Security/OTEL
  license:
    name: MIT + Non-Commercial
    url: https://github.com/hitechcloud-vietnam/ai-os/blob/main/LICENSE
  contact:
    name: HiTechCloud Vietnam
    url: https://github.com/hitechcloud-vietnam/ai-os

servers:
  - url: http://localhost:8080
    description: Local development (AI Gateway)
  - url: http://localhost:8081
    description: MCP Gateway (JSON-RPC 2.0)

security:
  - BearerAuth: []
  - ApiKeyAuth: []

paths:
  /health:
    get:
      summary: Health check
      operationId: healthCheck
      tags: [System]
      security: []
      responses:
        '200':
          description: Service healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /v1/registry/mcp-servers:
    get:
      summary: List MCP servers
      operationId: listMcpServers
      tags: [Registry - MCP]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of MCP servers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/McpServer'
    post:
      summary: Create MCP server
      operationId: createMcpServer
      tags: [Registry - MCP]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMcpServerRequest'
      responses:
        '201':
          description: MCP server created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpServer'

  /v1/registry/mcp-servers/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    get:
      summary: Get MCP server by ID
      operationId: getMcpServer
      tags: [Registry - MCP]
      responses:
        '200':
          description: MCP server details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpServer'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update MCP server
      operationId: updateMcpServer
      tags: [Registry - MCP]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMcpServerRequest'
      responses:
        '200':
          description: MCP server updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpServer'
    delete:
      summary: Delete MCP server
      operationId: deleteMcpServer
      tags: [Registry - MCP]
      responses:
        '204':
          description: MCP server deleted

  /v1/registry/skills:
    get:
      summary: List skills
      operationId: listSkills
      tags: [Registry - Skills]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Skill'
    post:
      summary: Create skill
      operationId: createSkill
      tags: [Registry - Skills]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSkillRequest'
      responses:
        '201':
          description: Skill created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'

  /v1/registry/skills/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    get:
      summary: Get skill by ID
      operationId: getSkill
      tags: [Registry - Skills]
      responses:
        '200':
          description: Skill details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update skill
      operationId: updateSkill
      tags: [Registry - Skills]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSkillRequest'
      responses:
        '200':
          description: Skill updated
    delete:
      summary: Delete skill
      operationId: deleteSkill
      tags: [Registry - Skills]
      responses:
        '204':
          description: Skill deleted

  /v1/registry/plugins:
    get:
      summary: List plugins
      operationId: listPlugins
      tags: [Registry - Plugins]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of plugins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Plugin'
    post:
      summary: Create plugin
      operationId: createPlugin
      tags: [Registry - Plugins]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePluginRequest'
      responses:
        '201':
          description: Plugin created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plugin'

  /v1/registry/plugins/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    get:
      summary: Get plugin by ID
      operationId: getPlugin
      tags: [Registry - Plugins]
      responses:
        '200':
          description: Plugin details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plugin'
    put:
      summary: Update plugin
      operationId: updatePlugin
      tags: [Registry - Plugins]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePluginRequest'
      responses:
        '200':
          description: Plugin updated
    delete:
      summary: Delete plugin
      operationId: deletePlugin
      tags: [Registry - Plugins]
      responses:
        '204':
          description: Plugin deleted

  /v1/registry/installations:
    get:
      summary: List installations
      operationId: listInstallations
      tags: [Registry - Installations]
      responses:
        '200':
          description: List of installations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Installation'
    post:
      summary: Create installation
      operationId: createInstallation
      tags: [Registry - Installations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInstallationRequest'
      responses:
        '201':
          description: Installation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Installation'

  /v1/registry/search:
    get:
      summary: Search across all entities
      operationId: search
      tags: [Registry - Search]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'

  /v1/admin/organizations:
    get:
      summary: List organizations (admin)
      operationId: listOrganizations
      tags: [Admin]
      responses:
        '200':
          description: List of organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Organization'

  /v1/admin/organizations/{id}/members:
    get:
      summary: Get organization members (admin)
      operationId: getOrganizationMembers
      tags: [Admin]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Organization members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'

  /v1/admin/audit-logs:
    get:
      summary: Get audit logs (admin)
      operationId: getAuditLogs
      tags: [Admin]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Audit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditLog'

  /mcp/v1:
    post:
      summary: MCP JSON-RPC 2.0 endpoint
      operationId: mcpJsonRpc
      tags: [MCP Gateway]
      description: |
        Send JSON-RPC 2.0 requests to the MCP Gateway.
        Supported methods: initialize, tools/list, tools/call, resources/list, prompts/list, ping
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonRpcRequest'
      responses:
        '200':
          description: JSON-RPC response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonRpcResponse'

  /mcp/v1/sse:
    get:
      summary: MCP SSE stream
      operationId: mcpSse
      tags: [MCP Gateway]
      description: Server-Sent Events stream for real-time MCP notifications
      responses:
        '200':
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string

  /tools:
    get:
      summary: List all available tools
      operationId: listTools
      tags: [MCP Gateway]
      responses:
        '200':
          description: All aggregated MCP tools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/McpTool'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0

  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
        version:
          type: string
          example: "0.1.0"
        service:
          type: string

    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        plan:
          type: string
          enum: [free, pro, enterprise]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
        role:
          type: string
          enum: [owner, admin, member, viewer]
        created_at:
          type: string
          format: date-time

    McpServer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        transport:
          type: string
          enum: [stdio, sse, streamable-http]
        endpoint_url:
          type: string
          format: uri
        capabilities:
          type: object
        verified:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    Skill:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        current_version:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    Plugin:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        current_version:
          type: string
        plugin_type:
          type: string
          enum: [mcp-server, api, webhook, custom]
        verified:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    Installation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        installable_type:
          type: string
          enum: [mcp_server, skill, plugin]
        installable_id:
          type: string
          format: uuid
        config:
          type: object
        status:
          type: string
          enum: [active, paused, removed]
        created_at:
          type: string
          format: date-time

    AuditLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        actor_id:
          type: string
          format: uuid
        action:
          type: string
        resource_type:
          type: string
        resource_id:
          type: string
        metadata:
          type: object
        created_at:
          type: string
          format: date-time

    ApiKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        name:
          type: string
        key_prefix:
          type: string
        scopes:
          type: array
          items:
            type: string
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time

    McpTool:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        inputSchema:
          type: object
        server_name:
          type: string

    JsonRpcRequest:
      type: object
      required: [jsonrpc, method]
      properties:
        jsonrpc:
          type: string
          enum: ["2.0"]
        method:
          type: string
        params:
          type: object
        id:
          oneOf:
            - type: string
            - type: integer
            - type: 'null'

    JsonRpcResponse:
      type: object
      properties:
        jsonrpc:
          type: string
          enum: ["2.0"]
        result:
          type: object
        error:
          $ref: '#/components/schemas/JsonRpcError'
        id:
          oneOf:
            - type: string
            - type: integer
            - type: 'null'

    JsonRpcError:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object

    CreateMcpServerRequest:
      type: object
      required: [name, transport]
      properties:
        name:
          type: string
        description:
          type: string
        transport:
          type: string
          enum: [stdio, sse, streamable-http]
        endpoint_url:
          type: string
          format: uri
        capabilities:
          type: object

    UpdateMcpServerRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        endpoint_url:
          type: string
          format: uri
        capabilities:
          type: object

    CreateSkillRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string

    UpdateSkillRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string

    CreatePluginRequest:
      type: object
      required: [name, plugin_type]
      properties:
        name:
          type: string
        description:
          type: string
        plugin_type:
          type: string
          enum: [mcp-server, api, webhook, custom]

    UpdatePluginRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string

    CreateInstallationRequest:
      type: object
      required: [installable_type, installable_id]
      properties:
        installable_type:
          type: string
          enum: [mcp_server, skill, plugin]
        installable_id:
          type: string
          format: uuid
        config:
          type: object

    PaginatedResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    ErrorResponse:
      type: object
      properties:
        error:
          type: string